heap tester: all blocks ui refactor
