+++ FRI.AUS2.StuctureTester/FRI.AUS2.StructureTester.HeapFile/Controls/HeapFileBlocks.xaml
@@ -1,0 +1,19 @@
+﻿<UserControl x:Class="FRI.AUS2.StructureTester.HeapFileTester.Controls.HeapFileBlocks"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:local="clr-namespace:FRI.AUS2.StructureTester.HeapFileTester.Controls"
+    mc:Ignorable="d"
+    d:DesignHeight="450" d:DesignWidth="800">
+    <ScrollViewer>
+        <TreeView x:Name="_treeView_Blocks">
+            <TreeView.ItemContainerStyle>
+                <Style TargetType="{x:Type TreeViewItem}">
+                    <EventSetter Event="MouseDoubleClick"
+                        Handler="_treeView_Blocks_OnItemDoubleClicked" />
+                </Style>
+            </TreeView.ItemContainerStyle>
+        </TreeView>
+    </ScrollViewer>
+</UserControl>
