+++ FRI.AUS2.StuctureTester/FRI.AUS2.StructureTester.HeapFile/FRI.AUS2.StructureTester.HeapFileTester.csproj.user
@@ -10,6 +10,9 @@
     <Compile Update="Controls\HeapDataForm.xaml.cs">
       <SubType>Code</SubType>
     </Compile>
+    <Compile Update="Controls\HeapFileBlocks.xaml.cs">
+      <SubType>Code</SubType>
+    </Compile>
     <Compile Update="Controls\HeapFileStats.xaml.cs">
       <SubType>Code</SubType>
     </Compile>
@@ -19,6 +22,9 @@
   </ItemGroup>
   <ItemGroup>
     <Page Update="Controls\HeapDataForm.xaml">
+      <SubType>Designer</SubType>
+    </Page>
+    <Page Update="Controls\HeapFileBlocks.xaml">
       <SubType>Designer</SubType>
     </Page>
     <Page Update="Controls\HeapFileStats.xaml">
